% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_have_na.R
\name{cols_have_na}
\alias{cols_have_na}
\alias{cols_have_na.matrix}
\title{Check if columns have missing values}
\usage{
cols_have_na(.data, ...)

\method{cols_have_na}{matrix}(.data, condition = "none", ...)
}
\arguments{
\item{.data}{A two-dimensional data structure.}

\item{...}{Arguments passed on to \code{op_ctrl}
\describe{
  \item{output_class}{One of ("vector", "list"). Possibly abbreviated.}
  \item{cols}{A vector indicating which columns to consider for the operation. If \code{NULL}, all
columns are used. If its length is 0, no columns are considered. Negative values are supported.}
  \item{rows}{Like \code{cols} but for row indices.}
}}

\item{condition}{One of ("all", "any", "none"). Possibly abbreviated.}
}
\description{
Check if columns have missing values
}
