% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ctrl.R
\name{op_ctrl}
\alias{op_ctrl}
\title{Operation's control parameters}
\usage{
op_ctrl(data_class, output_mode, na_action = "exclude", ...)
}
\arguments{
\item{data_class}{See \code{\link[base:class]{base::class()}}.}

\item{output_mode}{Desired \code{\link[base:storage.mode]{base::storage.mode()}} for the result.}

\item{na_action}{One of "exclude" or "pass".}

\item{...}{Internal.}
}
\description{
Operation's control parameters
}
