% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_finites.R
\name{row_finites}
\alias{row_finites}
\alias{row_finites.matrix}
\alias{row_finites.data.frame}
\title{Conditions related to finite values}
\usage{
row_finites(.data, match_type = "none", ...)

\method{row_finites}{matrix}(.data, match_type = "none", ...)

\method{row_finites}{data.frame}(.data, match_type = "none", ...)
}
\arguments{
\item{.data}{A two-dimensional data structure.}

\item{match_type}{One of ("all", "any", "none", "which_first", "count"). Possibly abbreviated.}

\item{...}{Arguments passed on to \code{op_ctrl}
\describe{
  \item{output_class}{One of ("vector", "list", "data.frame", "matrix"), possibly abbreviated.}
  \item{cols}{A vector indicating which columns to consider for the operation. If \code{NULL}, all
columns are used. If its length is 0, no columns are considered. Negative numbers, logical
values, and character vectors representing column names are supported.}
  \item{rows}{Like \code{cols} but for row indices.}
}}
}
\description{
Conditions related to finite values
}
