% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_infs.R
\name{row_infs}
\alias{row_infs}
\alias{row_infs.matrix}
\alias{row_infs.data.frame}
\alias{row_infs.data.table}
\title{Check if a row's columns have infinite values}
\usage{
row_infs(.data, match_type = "none", ...)

\method{row_infs}{matrix}(.data, match_type = "none", ...)

\method{row_infs}{data.frame}(.data, match_type = "none", ...)

\method{row_infs}{data.table}(.data, ...)
}
\arguments{
\item{.data}{A two-dimensional data structure.}

\item{match_type}{One of ("all", "any", "none", "which_first"). Possibly abbreviated.}

\item{...}{Arguments passed on to \code{op_ctrl}
\describe{
  \item{output_class}{One of ("vector", "list", "data.frame", "matrix"), possibly abbreviated.}
  \item{cols}{A vector indicating which columns to consider for the operation. If \code{NULL}, all
columns are used. If its length is 0, no columns are considered. Negative numbers, logical
values, and character vectors representing column names are supported.}
  \item{rows}{Like \code{cols} but for row indices.}
}}
}
\description{
For each desired row, check if all/any/none of the columns have infinite values (\code{Inf}), or
return the index of the first column containing \code{Inf}.
}
\details{
Note that \code{match_type = "which_first"} will result in an integer output, whereas the other
options result in a logical output.
}
