% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{wiserow-package}
\alias{wiserow}
\alias{wiserow-package}
\title{Multi-threaded, coercion-free implementations of common row-wise operations}
\description{
Fast row-oriented operations implemented in C++, all of which are multi-threaded by leveraging
\pkg{RcppParallel} and \pkg{RcppThread}. Virtually no deep copies of input data are made, even of
character data thanks to the 'string_ref' class from the C++ 'Boost' library. In contrast to
other functions, the ones in this package support data frames with differently typed columns as
input without coercing to a matrix, performing, if necessary, on-the-fly type promotion according
to R rules (like transforming logicals to integers to allow summation).
}
\details{
Start by looking at \code{\link[=op_ctrl]{op_ctrl()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://asardaes.github.io/wiserow/}
  \item \url{https://github.com/asardaes/wiserow}
  \item Report bugs at \url{https://github.com/asardaes/wiserow/issues}
}

}
\author{
Alexis Sarda-Espinosa
}
