% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ctrl.R
\name{op_ctrl}
\alias{op_ctrl}
\title{Operation's control parameters}
\usage{
op_ctrl(output_mode, output_class = "vector", na_action = "exclude",
  cols = NULL, rows = NULL, ...)
}
\arguments{
\item{output_mode}{Desired \code{\link[base:storage.mode]{base::storage.mode()}} for the result.}

\item{output_class}{Only "vector" supported for now.}

\item{na_action}{One of "exclude" or "pass".}

\item{cols}{A vector indicating which columns to consider for the operation. If \code{NULL}, all
columns are used. If its length is 0, no columns are considered. Negative values are supported.}

\item{rows}{Like \code{cols} but for row indices.}

\item{...}{Internal.}
}
\description{
Operation's control parameters
}
