% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_arith.R
\name{row_arith}
\alias{row_arith}
\alias{row_arith.matrix}
\alias{row_arith.data.frame}
\alias{row_arith.data.table}
\title{Row-wise arithmetic operations}
\usage{
row_arith(.data, ...)

\method{row_arith}{matrix}(.data, operator = c("+", "-", "*", "/"),
  output_mode, ...)

\method{row_arith}{data.frame}(.data, operator = c("+", "-", "*", "/"),
  output_mode, ...)

\method{row_arith}{data.table}(.data, ...)
}
\arguments{
\item{.data}{A two-dimensional data structure.}

\item{...}{Arguments passed on to \code{op_ctrl}
\describe{
  \item{output_class}{One of ("vector", "list"), possibly abbreviated.}
  \item{na_action}{One of ("exclude", "pass"), possibly abbreviated.}
  \item{cols}{A vector indicating which columns to consider for the operation. If \code{NULL}, all
columns are used. If its length is 0, no columns are considered. Negative numbers, logical
values, and character vectors representing column names are supported.}
  \item{rows}{Like \code{cols} but for row indices.}
}}

\item{operator}{One of ("+", "-", "*", "/").}

\item{output_mode}{Output's \code{\link[base:storage.mode]{base::storage.mode()}}. If missing, it will be inferred.}
}
\description{
Row-wise arithmetic operations
}
\details{
Conceptually, this function takes each row, say \code{x}, and executes \code{Reduce(operator, unlist(x))}.
\code{x} may be a list of values with different non-character types (like a \code{data.frame} row), and
\code{NA} values can be excluded.
}
