% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_infs.R
\name{row_infs}
\alias{row_infs}
\alias{row_infs.matrix}
\alias{row_infs.data.frame}
\alias{row_infs.data.table}
\title{Check if a row's columns have infinite values}
\usage{
row_infs(.data, which_cols = "none", ...)

\method{row_infs}{matrix}(.data, which_cols = "none", ...)

\method{row_infs}{data.frame}(.data, which_cols = "none", ...)

\method{row_infs}{data.table}(.data, ...)
}
\arguments{
\item{.data}{A two-dimensional data structure.}

\item{which_cols}{One of ("all", "any", "none"). Possibly abbreviated.}

\item{...}{Arguments passed on to \code{op_ctrl}
\describe{
  \item{output_class}{One of ("vector", "list"). Possibly abbreviated.}
  \item{cols}{A vector indicating which columns to consider for the operation. If \code{NULL}, all
columns are used. If its length is 0, no columns are considered. Negative and logical values
are supported.}
  \item{rows}{Like \code{cols} but for row indices.}
}}
}
\description{
For each desired row, check if all/any/none of the columns have infinite values (\code{Inf}).
}
