% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_max.R
\name{row_max}
\alias{row_max}
\alias{row_max.matrix}
\alias{row_max.data.frame}
\title{Row-wise maxima}
\usage{
row_max(.data, ...)

\method{row_max}{matrix}(.data, ...)

\method{row_max}{data.frame}(.data, ...)
}
\arguments{
\item{.data}{A two-dimensional data structure.}

\item{...}{
  Arguments passed on to \code{\link[=op_ctrl]{op_ctrl}}
  \describe{
    \item{\code{output_class}}{One of ("vector", "list", "data.frame", "matrix"), possibly abbreviated.}
    \item{\code{na_action}}{One of ("exclude", "pass"), possibly abbreviated.}
    \item{\code{cols}}{A vector indicating which columns to consider for the operation. If \code{NULL}, all
columns are used. If its length is 0, no columns are considered. Negative numbers, logical
values, character vectors representing column names, and \link[tidyselect:select_helpers]{tidyselect::select_helpers} are
supported.}
    \item{\code{rows}}{Like \code{cols} but for row indices, and without \code{tidyselect} support.}
    \item{\code{factor_mode}}{One of ("character", "integer"), possibly abbreviated. If a column is a
factor, this determines whether the operation uses its internal integer values, or the
character values from its levels.}
  }}
}
\description{
Row-wise maxima
}
\details{
For data frames, if the input columns have different modes, all of them will be promoted
on-the-fly to the highest one in the R hierarchy.
}
\note{
String comparison is done in C++, which can vary from what R does.
}
