% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_nas.R
\name{row_nas}
\alias{row_nas}
\alias{row_nas.matrix}
\alias{row_nas.data.frame}
\title{Check if a row's columns have missing values}
\usage{
row_nas(.data, which_cols = "none", ...)

\method{row_nas}{matrix}(.data, which_cols = "none", ...)

\method{row_nas}{data.frame}(.data, which_cols = "none", ...)
}
\arguments{
\item{.data}{A two-dimensional data structure.}

\item{which_cols}{One of ("all", "any", "none"). Possibly abbreviated.}

\item{...}{Arguments passed on to \code{op_ctrl}
\describe{
  \item{output_class}{One of ("vector", "list"). Possibly abbreviated.}
  \item{cols}{A vector indicating which columns to consider for the operation. If \code{NULL}, all
columns are used. If its length is 0, no columns are considered. Negative and logical values
are supported.}
  \item{rows}{Like \code{cols} but for row indices.}
}}
}
\description{
For each desired row, check if all/any/none of the columns have missing values (\code{NA}).
}
