% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_sums.R
\name{row_sums}
\alias{row_sums}
\alias{row_sums.matrix}
\title{Row-wise sum}
\usage{
row_sums(.data, ...)

\method{row_sums}{matrix}(.data, output_mode = storage.mode(.data), ...)
}
\arguments{
\item{.data}{A two-dimensional data structure.}

\item{...}{Currently ignored.}

\item{output_mode}{Output's \code{\link[base:storage.mode]{base::storage.mode()}}.}
}
\description{
Row-wise sum
}
