% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_sums.R
\name{row_sums}
\alias{row_sums}
\alias{row_sums.matrix}
\title{Row-wise sum}
\usage{
row_sums(.data, ...)

\method{row_sums}{matrix}(.data, output_mode, ...)
}
\arguments{
\item{.data}{A two-dimensional data structure.}

\item{...}{Arguments passed on to \code{op_ctrl}
\describe{
  \item{output_class}{One of ("vector", "list"). Possibly abbreviated.}
  \item{na_action}{One of ("exclude", "pass"). Possibly abbreviated.}
  \item{cols}{A vector indicating which columns to consider for the operation. If \code{NULL}, all
columns are used. If its length is 0, no columns are considered. Negative values are supported.}
  \item{rows}{Like \code{cols} but for row indices.}
}}

\item{output_mode}{Output's \code{\link[base:storage.mode]{base::storage.mode()}}. If missing, it will be inferred.}
}
\description{
Row-wise sum
}
