% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{wiserow-package}
\alias{wiserow}
\alias{wiserow-package}
\title{Fast, multi-threaded implementations of common row-wise operations}
\description{
TBD.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asardaes/wiserow}
  \item Report bugs at \url{https://github.com/asardaes/wiserow/issues}
}

}
\author{
Alexis Sarda-Espinosa
}
