% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{wiserow-package}
\alias{wiserow}
\alias{wiserow-package}
\title{Multi-threaded, coercion-free implementations of common row-wise operations}
\description{
Fast row-oriented operations implemented in C++, all of which are multi-threaded by leveraging
\pkg{RcppParallel} and \pkg{RcppThread}. Virtually no deep copies of input data are made, even of
character data thanks to Boost's \code{string_ref}. In contrast to other functions, the ones in this
package support data frames with differently typed columns as input without coercion to a matrix,
performing on-the-fly type promotion following R rules, where necessary.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asardaes/wiserow}
  \item Report bugs at \url{https://github.com/asardaes/wiserow/issues}
}

}
\author{
Alexis Sarda-Espinosa
}
